:shirt: refactor(views): Move json views to a separate file
