+++ app/views/antminer.py
@@ -5,6 +5,11 @@ from flask import (jsonify,
                    url_for,
                    flash,
                    )
+from flask.views import MethodView
+from app.views.antminer_json import (get_summary,
+                                     get_pools,
+                                     get_stats,
+                                     )
 from sqlalchemy.exc import IntegrityError
 from app.pycgminer import CgminerAPI
 from app import app, db, logger, __version__
@@ -12,45 +17,6 @@ from app.models import Miner, MinerModel, Settings
 import re
 from datetime import timedelta
 import time
-
-
-def get_summary(ip):
-    cgminer = CgminerAPI(host=ip)
-    output = cgminer.summary()
-    output.update({"IP": ip})
-    return dict(output)
-
-
-def get_pools(ip):
-    cgminer = CgminerAPI(host=ip)
-    output = cgminer.pools()
-    output.update({"IP": ip})
-    return dict(output)
-
-
-def get_stats(ip):
-    cgminer = CgminerAPI(host=ip)
-    output = cgminer.stats()
-    output.update({"IP": ip})
-    return dict(output)
-
-
-@app.route('/<ip>/summary')
-def summary(ip):
-    output = get_summary(ip)
-    return jsonify(output)
-
-
-@app.route('/<ip>/pools')
-def pools(ip):
-    output = get_pools(ip)
-    return jsonify(output)
-
-
-@app.route('/<ip>/stats')
-def stats(ip):
-    output = get_stats(ip)
-    return jsonify(output)
 
 
 @app.route('/')
